Move anisotropic diffusion tests to different folder. (#27280)
