+++ test/tests/linearfvkernels/diffusion/tests
@@ -1,5 +1,5 @@
 [Tests]
-  design = 'LinearFVDiffusion.md LinearFVAnisotropicDiffusion.md'
+  design = 'LinearFVDiffusion.md'
   issues = '#25722'
   [mms-diffusion-1d-cd]
     type = PythonUnitTest
@@ -33,33 +33,6 @@
     input = 'test.py'
     test_case = TestDiffusion2DRZ
     requirement = 'The system shall display second-order convergence for diffusion problems with a central differencing scheme solved using a linear finite volume system on an orthogonal mesh in cylindrical coordinate system.'
-    required_python_packages = 'pandas matplotlib'
-    installation_type = in_tree # see #26480
-    max_threads = 1 # see libmesh issue #3808
-  []
-  [mms-anisotropic-diffusion-2d-orthogonal]
-    type = PythonUnitTest
-    input = 'test.py'
-    test_case = TestAnisotropicDiffusion2DOrthogonal
-    requirement = 'The system shall display second-order convergence for anisotropic diffusion problems with a central differencing scheme solved using a linear finite volume system on a two-dimensional orthogonal mesh.'
-    required_python_packages = 'pandas matplotlib'
-    installation_type = in_tree # see #26480
-    max_threads = 1 # see libmesh issue #3808
-  []
-  [mms-anisotropic-diffusion-2d-nonorthogonal]
-    type = PythonUnitTest
-    input = 'test.py'
-    test_case = TestAnisotropicDiffusion2DNonorthogonal
-    requirement = 'The system shall display second-order convergence for anisotropic diffusion problems with a central differencing scheme solved using a linear finite volume system on a two-dimensional nonorthogonal mesh.'
-    required_python_packages = 'pandas matplotlib'
-    installation_type = in_tree # see #26480
-    max_threads = 1 # see libmesh issue #3808
-  []
-  [mms-anisotropic-diffusion-2d-rz]
-    type = PythonUnitTest
-    input = 'test.py'
-    test_case = TestAnisotropicDiffusion2DRZ
-    requirement = 'The system shall display second-order convergence for anisotropic diffusion problems with a central differencing scheme solved using a linear finite volume system on an orthogonal mesh in cylindrical coordinate system.'
     required_python_packages = 'pandas matplotlib'
     installation_type = in_tree # see #26480
     max_threads = 1 # see libmesh issue #3808
